fix(desktop): keep packaging hook inside workspace
